More little fix
